created admin permissions

+++ src/pages/api/card/delete.js
@@ -1,17 +1,23 @@
 import { deleteCardById } from "server/mongodb/actions/Card";
 import { withSessionRoute } from "src/utils/lib/session";
+import { getUserFromId } from "server/mongodb/actions/User";
 
 // @route   DELETE api/card/delete
 // @desc    Delete Card Request
 // @access  Public
 const handler = async (req, res) => {
   try {
-    // const userId = req.session.user.id;
-    await deleteCardById(req.body);
+    const userId = req.session.user.id;
+    const user = await getUserFromId(userId);
+    if (user.isAdmin) {
+      await deleteCardById(req.body);
 
-    return res.status(200).json({
-      success: true,
-    });
+      return res.status(200).json({
+        success: true,
+      });
+    } else {
+      throw new Error("You do not have permission to do this action!");
+    }
   } catch (error) {
     res.status(400).json({
       success: false,
